requirements.txt
attrs==19.3.0
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
decorator==4.4.1
Fiona==1.8.13
geopandas==0.6.2
haversine==2.1.2
munch==2.5.0
networkx==2.4
numpy==1.17.4
pandas==0.25.3
pyproj==2.4.2.post1
python-dateutil==2.8.1
pytz==2019.3
Shapely==1.6.4.post2
six==1.13.0
tqdm==4.40.1
kafka-python==2.0.1
